login implemented to navbar
