+++ bone-client/components/Navigation/Nav.tsx
@@ -7,6 +7,7 @@ import ButtonBlue from "../Button/ButtonBlue";
 import ButtonRed from "../Button/ButtonRed";
 import { useRouter } from "next/navigation";
 import { signOut, useSession } from "next-auth/react";
+import ButtonLogout from "../Button/ButtonLogout";
 
 interface Props {
   openNav: () => void;
@@ -18,29 +19,15 @@ const Nav = ({ openNav }: Props) => {
 
   const showSession = () => {
     if (status === "authenticated") {
-      return (
-        // <ButtonBlue text={"로그아웃"} link={"/"} />
-        <button
-          className="border border-solid border-black rounded"
-          onClick={() => {
-            signOut({ redirect: false }).then(() => {
-              router.push("/");
-            });
-          }}
-        >
-          Sign Out
-        </button>
-      );
+      return <ButtonLogout text={"로그아웃"} />;
     } else if (status === "loading") {
       return <span className="text-[#888] text-sm mt-7">Loading...</span>;
     } else {
       return (
-        <Link
-          href="/login"
-          className="border border-solid border-black rounded"
-        >
-          Sign In
-        </Link>
+        <>
+          <ButtonBlue link="/login" text="로그인" />
+          <ButtonRed link="/register" text="회원가입" />
+        </>
       );
     }
   };
@@ -75,11 +62,7 @@ const Nav = ({ openNav }: Props) => {
           </li>
         </ul>
         <div className="flex items-center space-x-2 md:space-x-5">
-          {/* 로그인 */}
           {showSession()}
-          <ButtonBlue link="/login" text="로그인" />
-          {/* 회원가입 */}
-          <ButtonRed link="/register" text="회원가입" />
           <Bars3BottomRightIcon
             onClick={openNav}
             className="w-[1.5rem] lg:hidden h-[1.5rem] text-slate-900 cursor-pointer"
